app/backend/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Azure SDKs
azure-identity>=1.15.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.19.0
azure-ai-projects>=1.0.0b5
openai>=1.12.0

# Microsoft Agent Framework
agent-framework-core>=1.0.0b260107
agent-framework-azure-ai>=1.0.0b260107

# Utilities
httpx>=0.26.0
pydantic>=2.5.0
pyyaml>=6.0.0
